+++ tests/testthat/test-fill_description.R
@@ -6,7 +6,7 @@ dir.create(dummypackage)
 fill_description(
   pkg = dummypackage,
   fields = list(
-    Title = "Build A Package From Rmarkdown file",
+    Title = "Build a package from Rmarkdown file",
     Description = "Use Rmarkdown First method to build your package. Start your package with documentation. Everything can be set from a Rmarkdown file in your project.",
     `Authors@R` = c(
       person("Sebastien", "Rochette", email = "sebastien@thinkr.fr", role = c("aut", "cre"), comment = c(ORCID = "0000-0002-1565-9313")),
@@ -26,6 +26,9 @@ test_that("fill_description adds DESCRIPTION", {
   ))
   lines <- readLines(file.path(dummypackage, "DESCRIPTION"))
   expect_true(lines[1] == "Package: dummypackage")
+
+  # Title Case changed
+  expect_equal(lines[3], "Title: Build A Package From Rmarkdown File")
 })
 
 # Fill description stops if malformed
@@ -51,3 +54,8 @@ test_that("no dot description fails", {
 
 # Delete dummy package
 unlink(dummypackage, recursive = TRUE)
+
+# Test capwords ----
+test_that("capwords works", {
+  expect_equal(capwords("using AIC for model selection"), "Using AIC For Model Selection")
+})